--- .gitignore
@@ -1,3 +1,0 @@
-/markov/
-*.json
-/.history/
