+++ go.mod
@@ -2,7 +2,6 @@ module go-markov
 
 go 1.18
 
-require (
-	github.com/jmcvetta/randutil v0.0.0-20150817122601-2bb1b664bcff // indirect
-	github.com/mroth/weightedrand v0.4.1 // indirect
-)
+require github.com/jmcvetta/randutil v0.0.0-20150817122601-2bb1b664bcff
+
+require github.com/stretchr/testify v1.8.0 // indirect
