+++ main.go
@@ -3,10 +3,12 @@ package main
 import (
 	"fmt"
 	"time"
+
+	"go-markov/markov"
 )
 
 func main() {
-	Start(StartInstructions{
+	markov.Start(markov.StartInstructions{
 		WriteMode:  "counter",
 		WriteLimit: 0,
 		StartKey:   "!S",
@@ -18,12 +20,12 @@ func main() {
 	//time.Sleep temporarily here because main exits faster than the goroutine
 	time.Sleep(1 * time.Second)
 
-	oi := OutputInstructions{
+	oi := markov.OutputInstructions{
 		Chain:  "test",
 		Method: "TargetedBeginning",
 		Target: "i",
 	}
-	output, err := Out(oi)
+	output, err := markov.Out(oi)
 
 	if err != nil {
 		panic(err)
