+++ markov/.gitignore
@@ -1,0 +1,4 @@
+/markov/chains/
+*.json
+/.history/
+main.go
