+++ markov/input.go
@@ -1,4 +1,4 @@
-package main
+package markov
 
 import (
 	"fmt"
